chore: split FileEditor into a separate chunk
This allows the Monaco editor component to be asynchronously loaded when needed.
